+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to insta and cargo-insta are documented here.
 - Added support for rendering some invisibles in diffs.  This now also
   should make sure that ANSI sequences in strings are no longer screwing
   up the terminal output. (#308)
+- Prevent inline snapshots to be used in loops. (#307)
 
 ## 1.21.2